@@ -4,26 +4,24 @@
   "private": true,
   "homepage": "http://BardAlexandrovskiy.github.io/all-in-one",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
-    "gsap": "^3.11.0",
-    "moment": "^2.29.1",
-    "moment-timezone": "^0.5.34",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-moment": "^1.1.1",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.0.2",
-    "react-scripts": "4.0.3",
-    "react-transition-group": "^4.4.2",
-    "redux": "^4.1.2",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.0",
-    "swiper": "^7.2.0",
-    "typescript": "^4.5.5"
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-regular-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "gsap": "^3.11.4",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.40",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-moment": "^1.1.3",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.0",
+    "react-scripts": "^5.0.1",
+    "react-transition-group": "^4.4.5",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.2",
+    "swiper": "^8.4.6"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -52,8 +50,10 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.3",
-    "node-sass": "^6.0.1",
-    "redux-logger": "^3.0.6"
+    "gh-pages": "^5.0.0",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-logger": "^3.0.6",
+    "sass": "^1.57.1",
+    "typescript": "^4.9.4"
   }
 }
