@@ -1,9 +1,18 @@
-import ReactDOM from "react-dom";
+import * as ReactDOMClient from "react-dom/client";
 import App from "./App";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import "./assets/styles/styles.scss";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById("root");
+
+// Create a root.
+const root = ReactDOMClient.createRoot(container);
+
+// Initial render: Render an element to the root.
+root.render(<App tab="home" />);
+
+// During an update, there's no need to pass the container again.
+root.render(<App tab="profile" />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
